deploy: drop Kubernetes 1.16, add 1.19
Kubernetes 1.19 has been released and the final Kubernetes 1.16 was
released beginning of September. This change will also allow us to
update to external-provisioner 2.0.0 across all supported Kubernetes
versions because it depends on the V1 CSINode and CSIDriver APIs that
were introduced in 1.17.

When bringing up Kubernetes 1.19, we have to use a different scheduler
config API because the one we were using before is no longer
supported. We also unconditionally enable the new alpha features
because we want to use them for PMEM-CSI.