Revert "pmem-csi-driver: Unregister node on exit"
This reverts commit acfeaf686dad8226bf77b30ff8121139641c0645.

As suspected during code review at the time, this change is
problematic because when the controller exits first, the nodes cannot
unregister. gRPC then tries to connect for up to 20 seconds before
eventually giving up (seen in CI testing of
https://github.com/intel/pmem-csi/pull/723). When reinstalling the
driver a lot as in the version skew testing, that tends to add up.

Relying on graceful unregistration also doesn't help in other cases. A
better solution is to detect dead nodes on the controller side.