WIP: dump information about VolumeAttachments
We shouldn't have any of those during PMEM-CSI testing because we use
skip attach. However, a recent failure in external-provisioner during
downgrade testing indicates that there must have been a
VolumeAttachement object because external-provisioner refused to
delete the volume with a "still attached" error (from
https://github.com/kubernetes-csi/external-provisioner/pull/438/files#r488452875).

The purpose of the additional debug log is to confirm that
VolumeAttachement objects were erroneously created and what they
contain.