operator: external-provisioner 2.0, validate content of RBAC rules
When the YAML files were updated for Kubernetes 1.19,
external-provisioner was updated to 2.0.0 without also updating the
operator default. We don't have a test case for "operator default
matches YAML default" because the operator defaults always overwrite
the YAML values; this is hard to change, so we'll have to catch this
through code review.

What could have been caught was that the RBAC rules got out-of-sync,
except that object comparison was unnecessarily limited to just the
"spec" field of those objects which have it. Now almost all fields
are compared.

This highlighted that the RBAC rules in the operator were slightly
different than the ones from the reference YAML files. Now they are
identical.