scheduler: pass only the node names, fix documentation
With NodeCacheCapable=true in the Extender configuration, just the
node names are passed as arguments and expected in the results. As we
don't need more than that, that mode is better because it is more
efficient. Logging gets streamlined for that mode.

The install instructions lacked documentation for Kubernetes 1.19 and
for setting up /var/lib/scheduler/scheduler-config.yaml.