pmem-csi-driver: better TLS parameter description and defaults
When using the same file names as under test/pmem-ca as defaults it
becomes a bit more obvious which file is what.

The description of the client files was misleading because it wasn't
clear that `certFile` and `certKey` are not file names, but rather the
names of the other parameters.